requirements.txt
flask
flask_cors
sqlalchemy
pydantic==2.11.3
phonenumbers
email_validator
imagekitio
python-dotenv